smartspend/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
tzdata>=2024.1
